+++ src/components/ContactForm/index.js
@@ -1,0 +1,3 @@
+import { ContactForm } from "./ContactForm.component";
+
+export { ContactForm };