+++ src/components/RequestStatusIndicator/index.js
@@ -1,0 +1,3 @@
+import { RequestStatusIndicator } from "./RequestStatusIndicator.component";
+
+export { RequestStatusIndicator };